# asset-log/docs/openapi.json
{
  "components": {
    "schemas": {
      "AccountResponse": {
        "properties": {
          "account_type": {
            "$ref": "#/components/schemas/AccountType"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "institution": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "withholding": {
            "type": [
              "boolean",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "account_type",
          "currency",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
...
      "AccountType": {
        "enum": [
          "tokutei",
          "ippan",
          "nisa_tsumitate",
          "nisa_growth",
          "ideco",
          "bank"
        ],
        "type": "string"
      },
      "AllocationItem": {
        "description": "レスポンスの1項目",
        "properties": {
          "key": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "ratio": {
            "description": "構成比（パーセント）。合計は必ず 100.00 になる",
            "example": "33.34",
            "type": "string"
          },
          "value_jpy": {
            "example": "500000",
            "type": "string"
          }
        },
        "required": [
          "key",
          "label",
          "value_jpy",
          "ratio"
        ],
        "type": "object"
      },
      "AllocationResult": {
        "properties": {
          "as_of": {
            "format": "date",
            "type": "string"
          },
          "base_currency": {
            "description": "常に \"JPY\"",
            "type": "string"
          },
          "fx_stale": {
            "type": "boolean"
          },
          "group_by": {
            "$ref": "#/components/schemas/GroupBy"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AllocationItem"
            },
            "type": "array"
          },
          "scope": {
            "description": "売買のある銘柄の評価額のみが対象。現金・預金残高は含まない。",
            "example": "securities_only",
            "type": "string"
          },
          "total_value_jpy": {
            "type": "string"
          },
          "unpriced_asset_count": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "as_of",
          "base_currency",
          "group_by",
          "scope",
          "fx_stale",
          "total_value_jpy",
          "unpriced_asset_count",
          "items"
        ],
        "type": "object"
      },
...
      "CreateAccountRequest": {
        "properties": {
          "account_type": {
            "$ref": "#/components/schemas/AccountType"
          },
          "currency": {
            "type": "string"
          },
          "institution": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "withholding": {
            "type": [
              "boolean",
              "null"
            ]
          }
        },
        "required": [
          "name",
          "account_type"
        ],
        "type": "object"
      },
...
      "FieldError": {
        "properties": {
          "field": {
            "description": "エラーが発生した入力項目名",
            "example": "quantity",
            "type": "string"
          },
          "message": {
            "description": "項目ごとのエラー内容",
            "example": "数量は正の数である必要があります",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "type": "object"
      },
...
      "Granularity": {
        "enum": [
          "day",
          "month"
        ],
        "type": "string"
      },
      "GroupBy": {
        "enum": [
          "none",
          "account_type",
          "asset_class",
          "account",
          "asset"
        ],
        "type": "string"
      },
...
      "HistoryPoint": {
        "properties": {
          "cost_jpy": {
            "example": "1000000",
            "type": "string"
          },
          "date": {
            "format": "date",
            "type": "string"
          },
          "market_value_jpy": {
            "example": "1234567",
            "type": "string"
          },
          "unpriced_asset_count": {
            "description": "その日、価格または約定日レートが引けず評価から外した銘柄数",
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "date",
          "market_value_jpy",
          "cost_jpy",
          "unpriced_asset_count"
        ],
        "type": "object"
      },
      "HistoryResult": {
        "properties": {
          "base_currency": {
            "description": "常に \"JPY\"",
            "type": "string"
          },
          "fx_stale": {
            "description": "為替の補充中に外部APIへ到達できず、キャッシュでしのいだ",
            "type": "boolean"
          },
          "granularity": {
            "$ref": "#/components/schemas/Granularity"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/HistorySeries"
            },
            "type": "array"
          },
          "source": {
            "$ref": "#/components/schemas/HistorySource"
          }
        },
        "required": [
          "granularity",
          "base_currency",
          "fx_stale",
          "series",
          "source"
        ],
        "type": "object"
      },
      "HistorySeries": {
        "properties": {
          "key": {
            "description": "分類軸のキー。group_by=none なら \"total\"、口座・銘柄軸ではUUID",
            "type": "string"
          },
          "label": {
            "description": "画面表示用の名前。enum 軸では日本語名、口座・銘柄軸では登録名。",
            "type": "string"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/HistoryPoint"
            },
            "type": "array"
          }
        },
        "required": [
          "key",
          "label",
          "points"
        ],
        "type": "object"
      },
      "HistorySource": {
        "description": "asset_history の評価結果がどちらの経路から来たか。",
        "enum": [
          "computed",
          "snapshot"
        ],
        "type": "string"
      },
...
      "ProblemDetails": {
        "description": "RFC 9457 準拠のエラーレスポンス（`application/problem+json`）",
        "properties": {
          "detail": {
            "description": "人間向けの詳細メッセージ",
            "example": "口座が見つかりません",
            "type": "string"
          },
          "errors": {
            "description": "項目単位のエラー（422のみ。空の場合はフィールドごと省略される）",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array"
          },
          "status": {
            "description": "HTTPステータスコード",
            "example": 404,
            "format": "int32",
            "maximum": 599,
            "minimum": 100,
            "type": "integer"
          },
          "title": {
            "description": "エラー種別の短い説明",
            "example": "Not Found",
            "type": "string"
          },
          "trace_id": {
            "description": "サーバーログと突き合わせるための識別子",
            "format": "uuid",
            "type": "string"
          },
          "type": {
            "description": "エラー種別の安定した識別子。クライアントはこの値で分岐できる",
            "example": "/errors/not-found",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "trace_id"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
      "UpdateAccountRequest": {
        "properties": {
          "institution": {
            "description": "金融機関名。未指定なら変更せず、`null` を明示すると削除する",
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "description": "口座名。未指定なら変更しない",
            "type": [
              "string",
              "null"
            ]
          },
          "withholding": {
            "description": "源泉徴収区分。未指定なら変更せず、`null` を明示すると削除する",
            "type": [
              "boolean",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "jobToken": {
        "description": "バッチ実行用の固定トークン",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "NISA/iDeCo を含む資産・取引の管理API",
    "license": {
      "name": ""
    },
    "title": "asset-log API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/accounts": {
      "get": {
        "operationId": "list_accounts",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AccountResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "口座の一覧"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "認証が必要"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "accounts"
        ]
      },
      "post": {
        "operationId": "create_account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            },
            "description": "口座を作成した"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "認証が必要"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "同じ名前の口座が既に存在する"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "通貨コードの形式不正、口座名が空、源泉徴収区分の指定誤りなど"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "accounts"
        ]
      }
    },
    "/accounts/{id}": {
      "delete": {
        "operationId": "delete_account",
        "parameters": [
          {
            "description": "口座ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "削除した"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "認証が必要"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "口座が存在しない"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "取引が紐づいているため削除できない"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "accounts"
        ]
      },
      "get": {
        "operationId": "get_account",
        "parameters": [
          {
            "description": "口座ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            },
            "description": "口座の詳細"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "認証が必要"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "口座が存在しない、または他ユーザーの口座"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "accounts"
        ]
      },
      "patch": {
        "operationId": "update_account",
        "parameters": [
          {
            "description": "口座ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            },
            "description": "更新後の口座"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "更新する項目が1つも指定されていない"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "認証が必要"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "口座が存在しない"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "同じ名前の口座が既に存在する"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "入力値が制約を満たしていない"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "accounts"
        ]
      }
    },
    "/analytics/allocation": {
      "get": {
        "operationId": "get_allocation",
        "parameters": [
          {
            "description": "基準日。既定は当日（JST）。未来日は不可",
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "format": "date",
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "description": "分類軸。既定は asset_class。none は指定できない",
            "in": "query",
            "name": "group_by",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/GroupBy"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllocationResult"
                }
              }
            },
            "description": "指定日時点のアセットアロケーション。ratio の合計は必ず100.00"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "認証が必要"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "未来日、group_by=none の指定"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "為替レートを取得できず、キャッシュにも無い"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/asset-history": {
      "get": {
        "operationId": "get_asset_history",
        "parameters": [
          {
            "description": "開始日。既定は to の365日前",
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "format": "date",
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "description": "終了日。既定は当日（JST）。未来日は不可",
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "format": "date",
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "description": "既定は day。期間が2000点を超える場合は month を使う",
            "in": "query",
            "name": "granularity",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Granularity"
            }
          },
          {
            "description": "分類軸。既定は none（全体合計の1系列）",
            "in": "query",
            "name": "group_by",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/GroupBy"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResult"
                }
              }
            },
            "description": "資産評価額の時系列。日次スナップショットが揃っていれば source=snapshot、なければ再計算して source=computed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "認証が必要"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "未来日、from が to より後、期間が2000点を超える"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "為替レートを取得できず、キャッシュにも無い"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "ローカル開発",
      "url": "http://localhost:8080"
    }
  ],
...
}